+++ Frontend/rujta-frontend/src/features/user/components/CartDrawer.jsx
@@ -1,26 +1,27 @@
 import React, { useState, useRef } from "react";
 import { IoMdClose } from "react-icons/io";
-import { FaPlus, FaMinus } from "react-icons/fa";
+import { FaPlus, FaMinus, FaPills } from "react-icons/fa";
+import { MdShoppingCartCheckout } from "react-icons/md";
+import { RiDeleteBin6Line } from "react-icons/ri";
 import { useNavigate } from "react-router-dom";
 import audio from "../../../assets/audio.wav";
 
 const CartDrawerUser = ({ cart, setCart, isOpen, onClose }) => {
   const navigate = useNavigate();
   const [isCheckingOut, setIsCheckingOut] = useState(false);
-
   const clickSound = useRef(new Audio(audio));
 
   const handleIncrease = (id) => {
-    setCart((prevCart) =>
-      prevCart.map((item) =>
+    setCart((prev) =>
+      prev.map((item) =>
         item.id === id ? { ...item, quantity: item.quantity + 1 } : item,
       ),
     );
   };
 
   const handleDecrease = (id) => {
-    setCart((prevCart) =>
-      prevCart
+    setCart((prev) =>
+      prev
         .map((item) =>
           item.id === id ? { ...item, quantity: item.quantity - 1 } : item,
         )
@@ -28,18 +29,19 @@ const CartDrawerUser = ({ cart, setCart, isOpen, onClose }) => {
     );
   };
 
-  const total = cart.reduce(
-    (sum, item) => sum + parseFloat(item.price) * item.quantity,
-    0,
-  );
+  const handleRemove = (id) => {
+    setCart((prev) => prev.filter((item) => item.id !== id));
+  };
+
+  const handleNavigateToMedicine = (id) => {
+    onClose();
+    navigate(`/user/medicine/${id}`);
+  };
 
   const handleCheckout = () => {
     if (isCheckingOut) return;
-
     setIsCheckingOut(true);
-
     clickSound.current.play();
-
     setTimeout(() => {
       onClose();
       navigate("/user/Checkout");
@@ -47,113 +49,207 @@ const CartDrawerUser = ({ cart, setCart, isOpen, onClose }) => {
   };
 
   return (
-    <div
-      className={`fixed right-0 top-0 h-full w-[320px] transform bg-white shadow-2xl transition-transform duration-300 dark:bg-gray-900 ${
-        isOpen ? "translate-x-0" : "translate-x-full"
-      } z-50 flex flex-col`}
-    >
-      {/* Header */}
-      <div className="flex items-center justify-between border-b border-gray-200 px-5 py-4 dark:border-gray-700">
-        {/* Left: Title + items count */}
-        <div className="flex items-baseline gap-2">
-          <h2 className="text-lg font-bold dark:text-white">Your Cart</h2>
-          {cart.length > 0 && (
-            <span className="text-xs text-gray-400 dark:text-gray-500">
-              {cart.length} {cart.length === 1 ? "item" : "items"}
-            </span>
-          )}
-        </div>
+    <>
+      {/* ── Backdrop ───────────────────────────────────────────── */}
+      {isOpen && (
+        <div
+          className="fixed inset-0 z-40 bg-black/20 backdrop-blur-sm"
+          onClick={onClose}
+        />
+      )}
 
-        {/* Right: clear all + close */}
-        <div className="flex items-center gap-3">
-          {cart.length > 0 && (
-            <button
-              onClick={() => setCart([])}
-              className="text-[11px] font-medium text-gray-400 underline-offset-2 transition-colors duration-200 hover:text-red-500 hover:underline dark:text-gray-500 dark:hover:text-red-400"
-            >
-              clear all
-            </button>
-          )}
-          <div className="h-4 w-px bg-gray-200 dark:bg-gray-700" />
-          <IoMdClose
-            className="cursor-pointer text-2xl text-gray-400 transition-colors hover:text-red-500"
-            onClick={onClose}
-          />
-        </div>
-      </div>
-
-      {/* Cart Items */}
-      <div className="flex-1 space-y-4 overflow-y-auto p-5">
-        {cart.length === 0 ? (
-          <p className="mt-10 text-center text-gray-500 dark:text-gray-300">
-            Your cart is empty 🛒
-          </p>
-        ) : (
-          cart.map((item) => (
-            <div
-              key={item.id}
-              className="flex items-center justify-between rounded-xl bg-gray-100 p-3 shadow-sm transition-all duration-200 hover:shadow-md dark:bg-gray-800"
-            >
-              <div>
-                <p className="font-semibold text-gray-800 dark:text-white">
-                  {item.name} ×{item.quantity}
-                </p>
-                <p className="text-sm text-gray-500 dark:text-gray-300">
-                  {(parseFloat(item.price) * item.quantity).toFixed(2)} EGP
-                </p>
-              </div>
-
-              <div className="flex items-center gap-2">
-                <button
-                  onClick={() => handleDecrease(item.id)}
-                  className="rounded-full bg-gray-200 p-1 hover:bg-gray-300 dark:bg-gray-700 dark:hover:bg-gray-600"
-                >
-                  <FaMinus className="text-xs text-gray-700 dark:text-gray-200" />
-                </button>
-                <span className="w-5 text-center text-sm font-semibold text-gray-800 dark:text-gray-200">
-                  {item.quantity}
-                </span>
-                <button
-                  onClick={() => handleIncrease(item.id)}
-                  className="rounded-full bg-gray-200 p-1 hover:bg-gray-300 dark:bg-gray-700 dark:hover:bg-gray-600"
-                >
-                  <FaPlus className="text-xs text-gray-700 dark:text-gray-200" />
-                </button>
-              </div>
+      {/* ── Drawer ─────────────────────────────────────────────── */}
+      <div
+        className={`fixed right-0 top-0 z-50 flex h-full w-[360px] flex-col bg-white shadow-2xl transition-transform duration-300 ease-in-out dark:bg-gray-900 ${
+          isOpen ? "translate-x-0" : "translate-x-full"
+        }`}
+      >
+        {/* ── Header ───────────────────────────────────────────── */}
+        <div className="flex items-center justify-between px-6 pb-4 pt-6">
+          <div className="flex items-center gap-3">
+            <div className="flex h-10 w-10 items-center justify-center rounded-2xl bg-secondary/10">
+              <MdShoppingCartCheckout className="text-xl text-secondary" />
             </div>
-          ))
-        )}
-      </div>
-
-      {/* Footer */}
-      {cart.length > 0 && (
-        <div className="sticky bottom-0 border-t border-gray-200 bg-white p-5 dark:border-gray-700 dark:bg-gray-900">
-          <div className="mb-3 flex items-center justify-between">
-            <span className="font-semibold text-gray-700 dark:text-gray-200">
-              Total:
-            </span>
-            <span className="font-bold text-gray-900 dark:text-white">
-              {total.toFixed(2)} EGP
-            </span>
+            <div>
+              <h2 className="text-lg font-bold leading-tight text-gray-800 dark:text-white">
+                Your Cart
+              </h2>
+              <p className="text-xs text-gray-400">
+                {cart.length === 0
+                  ? "No items"
+                  : `${cart.length} ${cart.length === 1 ? "item" : "items"}`}
+              </p>
+            </div>
           </div>
 
-          <button
-            onClick={handleCheckout}
-            disabled={isCheckingOut}
-            className={`flex w-full items-center justify-center gap-2 rounded-md bg-gradient-to-r from-secondary to-secondary py-2.5 font-semibold text-white transition-all duration-300 ${
-              isCheckingOut
-                ? "animate-pulse cursor-not-allowed opacity-80"
-                : "hover:scale-105 hover:opacity-90 active:scale-95"
-            }`}
-          >
-            {isCheckingOut && (
-              <div className="h-4 w-4 animate-spin rounded-full border-2 border-white border-t-transparent"></div>
+          <div className="flex items-center gap-2">
+            {cart.length > 0 && (
+              <button
+                onClick={() => setCart([])}
+                className="flex items-center gap-1.5 rounded-xl border border-red-100 bg-red-50 px-3 py-1.5 text-xs font-medium text-red-400 transition-all hover:bg-red-100 hover:text-red-600 dark:border-red-900/30 dark:bg-red-900/10"
+              >
+                <RiDeleteBin6Line className="text-sm" />
+                Clear
+              </button>
             )}
-            {isCheckingOut ? "Processing..." : "Checkout"}
-          </button>
+            <button
+              onClick={onClose}
+              className="flex h-9 w-9 items-center justify-center rounded-xl bg-gray-100 text-gray-500 transition-all hover:bg-gray-200 hover:text-gray-700 dark:bg-gray-800 dark:hover:bg-gray-700"
+            >
+              <IoMdClose className="text-lg" />
+            </button>
+          </div>
         </div>
-      )}
-    </div>
+
+        {/* ── Divider ──────────────────────────────────────────── */}
+        <div className="mx-6 h-px bg-gray-100 dark:bg-gray-800" />
+
+        {/* ── Items ────────────────────────────────────────────── */}
+        <div className="flex-1 overflow-y-auto px-5 py-4">
+          {cart.length === 0 ? (
+            <div className="flex flex-col items-center justify-center gap-4 pt-24">
+              <div className="flex h-24 w-24 items-center justify-center rounded-3xl bg-gray-50 text-5xl shadow-inner dark:bg-gray-800">
+                🛒
+              </div>
+              <div className="text-center">
+                <p className="font-bold text-gray-600 dark:text-gray-300">
+                  Cart is empty
+                </p>
+                <p className="mt-1 text-sm text-gray-400">
+                  Add some medicines to continue
+                </p>
+              </div>
+            </div>
+          ) : (
+            <div className="space-y-3">
+              {cart.map((item, index) => (
+                <div
+                  key={item.id}
+                  onClick={() => handleNavigateToMedicine(item.id)}
+                  className="relative cursor-pointer overflow-hidden rounded-2xl border border-gray-100 bg-gray-50/80 p-4 transition-all duration-200 hover:border-secondary/20 hover:bg-white hover:shadow-md dark:border-gray-700/50 dark:bg-gray-800/60"
+                  style={{ animationDelay: `${index * 50}ms` }}
+                >
+                  {/* ── ROW 1: Image + Name + Delete ─────────────── */}
+                  <div className="flex items-center gap-3">
+                    {/* ✅ Medicine Image with Fallback */}
+                    <div className="relative h-10 w-10 flex-shrink-0 overflow-hidden rounded-xl bg-gradient-to-br from-secondary/10 to-secondary/5 shadow-sm dark:bg-gray-700">
+                      {item.image || item.imageUrl ? (
+                        <img
+                          src={item.image || item.imageUrl}
+                          alt={item.name}
+                          className="h-full w-full object-cover"
+                          onError={(e) => {
+                            e.target.style.display = "none";
+                            e.target.nextSibling.style.display = "flex";
+                          }}
+                        />
+                      ) : null}
+                      {/* Fallback Icon */}
+                      <div
+                        className="absolute inset-0 flex items-center justify-center"
+                        style={{
+                          display:
+                            item.image || item.imageUrl ? "none" : "flex",
+                        }}
+                      >
+                        <FaPills className="text-lg text-secondary" />
+                      </div>
+                    </div>
+
+                    {/* Name */}
+                    <div className="flex-1 overflow-hidden">
+                      <p className="text-sm font-bold text-gray-800 dark:text-white">
+                        {item.name}
+                      </p>
+                    </div>
+
+                    {/* Delete */}
+                    <button
+                      onClick={(e) => {
+                        e.stopPropagation();
+                        handleRemove(item.id);
+                      }}
+                      className="flex h-8 w-8 flex-shrink-0 items-center justify-center rounded-xl bg-red-50 text-red-300 transition-all hover:bg-red-100 hover:text-red-500 dark:bg-red-900/10 dark:text-red-400"
+                    >
+                      <RiDeleteBin6Line className="text-base" />
+                    </button>
+                  </div>
+
+                  {/* ── ROW 2: Quantity + Controls ────────────────── */}
+                  <div
+                    className="mt-3 flex items-center justify-between"
+                    onClick={(e) => e.stopPropagation()}
+                  >
+                    {/* Quantity */}
+                    <div className="flex items-center gap-1.5">
+                      <div className="h-1.5 w-1.5 rounded-full bg-secondary" />
+                      <span className="text-xs font-semibold text-gray-400 dark:text-gray-500">
+                        {item.quantity} {item.quantity === 1 ? "unit" : "units"}
+                      </span>
+                    </div>
+
+                    {/* Controls */}
+                    <div className="flex items-center gap-1 rounded-2xl bg-white p-1 shadow-sm dark:bg-gray-700">
+                      <button
+                        onClick={() => handleDecrease(item.id)}
+                        className="flex h-7 w-7 items-center justify-center rounded-xl bg-gray-100 text-gray-500 transition-all hover:bg-red-100 hover:text-red-500 dark:bg-gray-600"
+                      >
+                        <FaMinus className="text-[9px]" />
+                      </button>
+                      <span className="w-7 text-center text-sm font-extrabold text-gray-800 dark:text-white">
+                        {item.quantity}
+                      </span>
+                      <button
+                        onClick={() => handleIncrease(item.id)}
+                        className="flex h-7 w-7 items-center justify-center rounded-xl bg-gray-100 text-gray-500 transition-all hover:bg-green-100 hover:text-green-500 dark:bg-gray-600"
+                      >
+                        <FaPlus className="text-[9px]" />
+                      </button>
+                    </div>
+                  </div>
+                </div>
+              ))}
+            </div>
+          )}
+        </div>
+
+        {/* ── Footer ───────────────────────────────────────────── */}
+        {cart.length > 0 && (
+          <div className="px-5 pb-6 pt-3">
+            <div className="mb-3 flex items-center justify-between rounded-2xl bg-secondary/5 px-4 py-3">
+              <span className="text-sm font-medium text-gray-500">
+                Total items
+              </span>
+              <span className="text-sm font-extrabold text-secondary">
+                {cart.reduce((sum, item) => sum + item.quantity, 0)} units
+              </span>
+            </div>
+
+            <button
+              onClick={handleCheckout}
+              disabled={isCheckingOut}
+              className={`flex w-full items-center justify-center gap-2.5 rounded-2xl bg-secondary py-3.5 text-sm font-bold text-white shadow-lg shadow-secondary/30 transition-all duration-300 ${
+                isCheckingOut
+                  ? "animate-pulse cursor-not-allowed opacity-75"
+                  : "hover:scale-[1.02] hover:shadow-xl hover:shadow-secondary/40 active:scale-[0.98]"
+              }`}
+            >
+              {isCheckingOut ? (
+                <>
+                  <div className="h-4 w-4 animate-spin rounded-full border-2 border-white border-t-transparent" />
+                  Processing...
+                </>
+              ) : (
+                <>
+                  <MdShoppingCartCheckout className="text-lg" />
+                  Proceed to Checkout
+                </>
+              )}
+            </button>
+          </div>
+        )}
+      </div>
+    </>
   );
 };
