+++ Frontend/rujta-frontend/src/features/user/components/checkout/MedicineChip.jsx
@@ -1,0 +1,43 @@
+import React from "react";
+
+const MedicineChip = ({ medicine, checked, onChange, disabled }) => {
+  const enough = medicine.isQuantityEnough;
+
+  return (
+    <label
+      className={`flex cursor-pointer select-none items-center gap-2 rounded-xl border px-3 py-2 transition-all duration-150 ${
+        disabled ? "cursor-not-allowed opacity-60" : "hover:shadow-sm"
+      } ${
+        checked && !disabled
+          ? "border-green-300 bg-green-50"
+          : enough
+            ? "border-gray-200 bg-white hover:border-green-200 hover:bg-green-50/50"
+            : "border-purple-100 bg-purple-50/40"
+      }`}
+    >
+      <input
+        type="checkbox"
+        checked={checked}
+        onChange={onChange}
+        disabled={disabled}
+        className="h-4 w-4 rounded border-gray-300 accent-green-500 focus:ring-green-400"
+      />
+      <span
+        className={`text-sm font-medium ${enough ? "text-gray-800" : "text-purple-700"}`}
+      >
+        {medicine.medicineName}
+      </span>
+      <span
+        className={`ml-auto rounded-full px-2 py-0.5 text-xs font-semibold ${
+          enough
+            ? "bg-green-100 text-green-700"
+            : "bg-purple-100 text-purple-600"
+        }`}
+      >
+        need {medicine.requestedQuantity}
+      </span>
+    </label>
+  );
+};
+
+export default MedicineChip;
