+++ Frontend/rujta-frontend/src/features/user/hooks/useCheckout.js
@@ -1,0 +1,366 @@
+import { useState, useEffect, useCallback } from "react";
+import { usePharmacies } from "../../pharmacies/hooks/usePharmacies";
+import { useOrders } from "../../orders/hooks/useOrders";
+import { useAuth } from "../../auth/hooks/useAuth";
+import useAddress from "../../address/hook/useAddress";
+import apiClient from "../../../shared/api/apiClient";
+import { decodePolyline } from "../../../utils/decodePolyline";
+
+export const useCheckout = () => {
+  const [cart, setCart] = useState([]);
+  const { pharmacies, loading, error, fetchPharmacies } = usePharmacies();
+  const { fetchUser } = useOrders();
+  const { user } = useAuth();
+  const {
+    addresses,
+    loading: addressesLoading,
+    error: addressesError,
+    fetchUserAddresses,
+    create: createAddress,
+    fetchById,
+  } = useAddress();
+
+  // ── Address States ──────────────────────────────────────────────
+  const [pharmaciesRange, setPharmaciesRange] = useState(5);
+  const [showLocationPrompt, setShowLocationPrompt] = useState(false);
+  const [showAddressSelection, setShowAddressSelection] = useState(true);
+  const [selectedAddressId, setSelectedAddressId] = useState(null);
+  const [showNewAddressForm, setShowNewAddressForm] = useState(false);
+  const [isConfirmingAddress, setIsConfirmingAddress] = useState(false);
+  const [newAddressForm, setNewAddressForm] = useState({
+    Street: "",
+    BuildingNo: "",
+    City: "",
+    Governorate: "",
+    IsDefault: false,
+  });
+
+  // ── Pharmacy / Order States ─────────────────────────────────────
+  const [expandedPharmacies, setExpandedPharmacies] = useState({});
+  const [showPaymentModal, setShowPaymentModal] = useState(false);
+  const [selectedPharmacyForPayment, setSelectedPharmacyForPayment] =
+    useState(null);
+  const [paymentMethod, setPaymentMethod] = useState("Cash");
+  const [selectedPharmacies, setSelectedPharmacies] = useState([]);
+  const [creatingOrder, setCreatingOrder] = useState(false);
+  const [selectedMedicines, setSelectedMedicines] = useState({});
+
+  // ── Map States ──────────────────────────────────────────────────
+  const [userLocation, setUserLocation] = useState(null);
+  const [deliveryAddressLocation, setDeliveryAddressLocation] = useState(null);
+  const [deliveryAddress, setDeliveryAddress] = useState(null);
+  const [hoveredPharmacyId, setHoveredPharmacyId] = useState(null);
+  const [routeData, setRouteData] = useState({});
+
+  // ── Toast State ─────────────────────────────────────────────────
+  const [toast, setToast] = useState(null);
+
+  const showToast = useCallback((type, message) => {
+    setToast({ type, message });
+    if (type === "success") setTimeout(() => setToast(null), 3200);
+  }, []);
+
+  // ── Route Fetching ──────────────────────────────────────────────
+  const fetchRoute = useCallback(
+    (pharmacy) => {
+      const start = deliveryAddressLocation || userLocation;
+      if (!start || !pharmacy) return;
+      const cacheKey = pharmacy.pharmacyId;
+
+      setRouteData((prev) => {
+        if (prev[cacheKey]) return prev;
+        const url =
+          `https://router.project-osrm.org/route/v1/driving/` +
+          `${start.lng},${start.lat};${pharmacy.longitude},${pharmacy.latitude}` +
+          `?overview=full&geometries=polyline`;
+        fetch(url)
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.code === "Ok" && data.routes?.[0]) {
+              const coordinates = decodePolyline(data.routes[0].geometry);
+              const distanceKm = (data.routes[0].distance / 1000).toFixed(2);
+              const durationMin = Math.round(data.routes[0].duration / 60);
+              setRouteData((p) => ({
+                ...p,
+                [cacheKey]: { coordinates, distanceKm, durationMin },
+              }));
+            }
+          })
+          .catch((err) =>
+            console.error("OSRM error:", pharmacy.pharmacyId, err),
+          );
+        return prev;
+      });
+    },
+    [deliveryAddressLocation, userLocation],
+  );
+
+  // ── Effects ─────────────────────────────────────────────────────
+  useEffect(() => {
+    if (pharmacies.length > 0) pharmacies.forEach(fetchRoute);
+  }, [pharmacies, fetchRoute]);
+
+  useEffect(() => {
+    if (!user) return;
+    const stored = JSON.parse(localStorage.getItem(`cart_${user.email}`)) || [];
+    setCart(stored);
+    fetchUserAddresses();
+  }, [user, fetchUserAddresses]);
+
+  useEffect(() => {
+    navigator.geolocation?.getCurrentPosition(
+      (pos) =>
+        setUserLocation({
+          lat: pos.coords.latitude,
+          lng: pos.coords.longitude,
+        }),
+      (err) => console.error("Geolocation error:", err),
+    );
+  }, []);
+
+  useEffect(() => {
+    const msg = typeof error === "string" ? error : error?.message || "";
+    if (msg.includes("location not set")) setShowLocationPrompt(true);
+  }, [error]);
+
+  // ── Handlers ────────────────────────────────────────────────────
+  const handleSetLocation = () => {
+    navigator.geolocation?.getCurrentPosition(async ({ coords }) => {
+      try {
+        await apiClient.put("/users/location", {
+          latitude: coords.latitude,
+          longitude: coords.longitude,
+        });
+        setUserLocation({ lat: coords.latitude, lng: coords.longitude });
+        setShowLocationPrompt(false);
+      } catch (err) {
+        console.error("Failed to update location:", err);
+      }
+    });
+  };
+
+  const handleNewAddressChange = (e) => {
+    const { name, value } = e.target;
+    setNewAddressForm((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleAddNewAddress = async () => {
+    try {
+      await createAddress(newAddressForm);
+      await fetchUserAddresses();
+      setShowNewAddressForm(false);
+    } catch (err) {
+      console.error("Failed to add address:", err);
+    }
+  };
+
+  const handleConfirmAddress = async () => {
+    if (!selectedAddressId) {
+      showToast("error", "Please select a delivery address!");
+      return;
+    }
+    setIsConfirmingAddress(true);
+    try {
+      if (cart.length > 0)
+        await fetchPharmacies(cart, selectedAddressId, pharmaciesRange);
+      const fullAddress = await fetchById(selectedAddressId);
+      if (fullAddress?.latitude && fullAddress?.longitude) {
+        setDeliveryAddressLocation({
+          lat: fullAddress.latitude,
+          lng: fullAddress.longitude,
+        });
+        setDeliveryAddress(fullAddress);
+      }
+      setShowAddressSelection(false);
+    } catch (err) {
+      console.error(err);
+    } finally {
+      setIsConfirmingAddress(false);
+    }
+  };
+
+  const handleExpandRange = async () => {
+    const newRange = pharmaciesRange + 5;
+    setPharmaciesRange(newRange);
+    if (selectedAddressId && cart.length > 0)
+      await fetchPharmacies(cart, selectedAddressId, newRange);
+  };
+
+  const handleTogglePharmacy = (pharmacyId) => {
+    if (selectedPharmacies.includes(pharmacyId)) {
+      setSelectedPharmacies((prev) => prev.filter((id) => id !== pharmacyId));
+      setSelectedMedicines((prev) => {
+        const u = { ...prev };
+        delete u[pharmacyId];
+        return u;
+      });
+    } else {
+      const pharmacy = pharmacies.find((p) => p.pharmacyId === pharmacyId);
+      const ids =
+        pharmacy?.foundMedicines
+          .filter((m) => m.isQuantityEnough)
+          .map((m) => m.medicineId) ?? [];
+      setSelectedPharmacies((prev) => [...prev, pharmacyId]);
+      setSelectedMedicines((prev) => ({ ...prev, [pharmacyId]: ids }));
+    }
+  };
+
+  const handleToggleMedicine = (pharmacyId, medicineId) => {
+    setSelectedMedicines((prev) => {
+      const meds = prev[pharmacyId] || [];
+      return {
+        ...prev,
+        [pharmacyId]: meds.includes(medicineId)
+          ? meds.filter((id) => id !== medicineId)
+          : [...meds, medicineId],
+      };
+    });
+  };
+
+  const handleOrderClick = (pharmacy) => {
+    setSelectedPharmacyForPayment(pharmacy);
+    setSelectedPharmacies([pharmacy.pharmacyId]);
+    setSelectedMedicines({
+      [pharmacy.pharmacyId]: pharmacy.foundMedicines
+        .filter((m) => m.isQuantityEnough)
+        .map((m) => m.medicineId),
+    });
+    setShowPaymentModal(true);
+  };
+
+  const handleConfirmOrders = async () => {
+    if (!cart.length) {
+      showToast("error", "Your cart is empty!");
+      return;
+    }
+    if (!selectedAddressId) {
+      showToast("error", "No delivery address selected!");
+      return;
+    }
+    if (!selectedPharmacies.length) {
+      showToast("error", "No pharmacies selected!");
+      return;
+    }
+
+    setCreatingOrder(true);
+    const orderDtos = selectedPharmacies.reduce((acc, pharmacyId) => {
+      const pharmacy = pharmacies.find((p) => p.pharmacyId === pharmacyId);
+      if (!pharmacy) return acc;
+      const items = pharmacy.foundMedicines.filter(
+        (m) =>
+          selectedMedicines[pharmacyId]?.includes(m.medicineId) &&
+          m.isQuantityEnough,
+      );
+      if (!items.length) return acc;
+      acc.push({
+        PharmacyID: pharmacyId,
+        DeliveryAddressId: selectedAddressId,
+        OrderItems: items.map((m) => ({
+          MedicineID: m.medicineId,
+          Quantity: m.requestedQuantity,
+        })),
+      });
+      return acc;
+    }, []);
+
+    if (!orderDtos.length) {
+      showToast(
+        "error",
+        "No valid orders. Check selected medicines & quantities.",
+      );
+      setCreatingOrder(false);
+      return;
+    }
+
+    try {
+      const { data: results } = await apiClient.post("/orders", orderDtos);
+      if (results?.length > 0) {
+        const orderedIds = new Set(
+          orderDtos.flatMap((d) => d.OrderItems.map((i) => i.MedicineID)),
+        );
+        const updatedCart = cart.filter((item) => !orderedIds.has(item.id));
+        localStorage.setItem(`cart_${user.email}`, JSON.stringify(updatedCart));
+        window.dispatchEvent(
+          new StorageEvent("storage", { key: `cart_${user.email}` }),
+        );
+        setCart(updatedCart);
+        setSelectedPharmacies([]);
+        setSelectedMedicines({});
+        await fetchUser();
+        showToast(
+          "success",
+          `${results.length} order${results.length > 1 ? "s" : ""} placed! 🎉`,
+        );
+        setTimeout(() => window.location.reload(), 3200);
+      } else {
+        showToast("error", "Failed to create orders. Please try again.");
+      }
+    } catch (err) {
+      console.error("Order error:", err);
+      showToast("error", "Failed to create orders. Check console for details.");
+    } finally {
+      setCreatingOrder(false);
+    }
+  };
+
+  const handlePaymentConfirm = async () => {
+    setShowPaymentModal(false);
+    if (paymentMethod === "Cash") await handleConfirmOrders();
+    else window.location.href = "/user/payment";
+  };
+
+  return {
+    // data
+    cart,
+    pharmacies,
+    loading,
+    error,
+    addresses,
+    addressesLoading,
+    addressesError,
+    // address states
+    pharmaciesRange,
+    showLocationPrompt,
+    showAddressSelection,
+    selectedAddressId,
+    showNewAddressForm,
+    isConfirmingAddress,
+    newAddressForm,
+    // pharmacy / order states
+    expandedPharmacies,
+    showPaymentModal,
+    selectedPharmacyForPayment,
+    paymentMethod,
+    selectedPharmacies,
+    creatingOrder,
+    selectedMedicines,
+    // map states
+    userLocation,
+    deliveryAddressLocation,
+    deliveryAddress,
+    hoveredPharmacyId,
+    routeData,
+    // toast
+    toast,
+    // setters
+    setSelectedAddressId,
+    setShowNewAddressForm,
+    setNewAddressForm,
+    setExpandedPharmacies,
+    setShowPaymentModal,
+    setPaymentMethod,
+    setHoveredPharmacyId,
+    setToast,
+    setShowAddressSelection,
+    // handlers
+    handleSetLocation,
+    handleNewAddressChange,
+    handleAddNewAddress,
+    handleConfirmAddress,
+    handleExpandRange,
+    handleTogglePharmacy,
+    handleToggleMedicine,
+    handleOrderClick,
+    handlePaymentConfirm,
+  };
+};
