+++ Frontend/rujta-frontend/src/utils/decodePolyline.js
@@ -1,0 +1,30 @@
+export const decodePolyline = (encoded) => {
+  let index = 0;
+  let lat = 0,
+    lng = 0;
+  const coordinates = [];
+
+  while (index < encoded.length) {
+    let shift = 0,
+      result = 0,
+      byte;
+    do {
+      byte = encoded.charCodeAt(index++) - 63;
+      result |= (byte & 0x1f) << shift;
+      shift += 5;
+    } while (byte >= 0x20);
+    lat += result & 1 ? ~(result >> 1) : result >> 1;
+
+    shift = 0;
+    result = 0;
+    do {
+      byte = encoded.charCodeAt(index++) - 63;
+      result |= (byte & 0x1f) << shift;
+      shift += 5;
+    } while (byte >= 0x20);
+    lng += result & 1 ? ~(result >> 1) : result >> 1;
+
+    coordinates.push([lat / 1e5, lng / 1e5]);
+  }
+  return coordinates;
+};
